requirements.txt
streamlit
scikit-learn
pandas
numpy
langgraph
langchain-core
langchain-groq
langchain-community
chromadb
sentence-transformers
python-dotenv
matplotlib
seaborn
torchvision
